Add useConfirm composable from CDH-Vue-Lib
