+++ frontend/components/layout/NavBar.vue
@@ -1,8 +1,63 @@
 <script lang="ts" setup>
-import { BSIcon } from "cdh-vue-lib";
+import { BSIcon, useConfirm } from "cdh-vue-lib";
 import useStaticFile from "~/composables/useStaticFile";
+import { useMutation } from "@vue/apollo-composable";
+import { graphql } from "~/generated/gql";
+import { useI18n } from "vue-i18n";
 
 const currentUser = computed(() => useCurrentUserStore().currentUser);
+
+const { t } = useI18n();
+
+const CREATE_STUDY_MUTATION = graphql(`
+    mutation CreateNewStudy {
+        createStudy {
+            study {
+                id
+                title
+            }
+            errors {
+                field
+                messages
+            }
+        }
+    }
+`);
+
+const { mutate: createStudy } = useMutation(CREATE_STUDY_MUTATION);
+
+const handleCreateStudy = async () => {
+    try {
+        const result = await createStudy();
+
+        if (!result?.data) {
+            useNotification("No response from server", "danger");
+            return;
+        }
+
+        if (result.data.createStudy?.errors?.length) {
+            useNotification("Failed to create study", "danger");
+            return;
+        }
+
+        if (result.data.createStudy?.study?.id) {
+            useNotification("Study created successfully", "success");
+            // Redirect to the study detail page
+            void navigateTo(`/studies/${result.data.createStudy.study.id}`);
+        }
+    } catch {
+        useNotification("Failed to create study. Please try again.", "danger");
+    }
+};
+
+function newRegistration(): void {
+    useConfirm({
+        text: t("Are you sure you want to start a new registration?"),
+        callback: () => {
+            void handleCreateStudy();
+        },
+    });
+}
 </script>
 
 <template>
@@ -36,13 +91,9 @@ const currentUser = computed(() => useCurrentUserStore().currentUser);
                         </NuxtLink>
                     </li>
                     <li>
-                        <NuxtLink
-                            v-if="currentUser"
-                            to="/procreg/start"
-                            class="nav-link"
-                            active-class="active"
-                            >{{ $t("New registration") }}
-                        </NuxtLink>
+                        <button class="nav-link" @click="newRegistration">
+                            {{ $t("New registration") }}
+                        </button>
                     </li>
                     <li v-if="currentUser">
                         <NuxtLink
